Ticket #48899 - Values of dbcachetries/dbcachehits in cn=monitor could overflow.
Description: The internal variable cache_tries is declared as u_int32_t.
The values provided from Berkeley DB are typed as uintmax_t, which has
64 bit width.  Since cache_tries stores the sum of st_cache_miss and
mpstat->st_cache_hit, the type should be also uintmax_t so that the
overflow problem is solved.

https://fedorahosted.org/389/ticket/48899

Reviewed by mreynolds@redhat.com (Thank you, Mark!!)

+++ ldap/servers/slapd/back-ldbm/monitor.c
@@ -226,7 +226,7 @@ int ldbm_back_monitor_search(Slapi_PBlock *pb, Slapi_Entry *e,
     char buf[BUFSIZ];
     DB_MPOOL_STAT *mpstat = NULL;
     DB_MPOOL_FSTAT **mpfstat = NULL;
-    u_int32_t cache_tries;
+    uintmax_t cache_tries;
 
     vals[0] = &val;
     vals[1] = NULL;
@@ -247,7 +247,7 @@ int ldbm_back_monitor_search(Slapi_PBlock *pb, Slapi_Entry *e,
 
     /* cache tries*/
     cache_tries = (mpstat->st_cache_miss + mpstat->st_cache_hit);
-    sprintf(buf, "%u", cache_tries);
+    sprintf(buf, "%lu", cache_tries);
     MSET("dbCacheTries");
 
     /* cache hit ratio*/
